backend/requirements.txt
# Utilities
pydantic
python-multipart
tiktoken
fastapi
fastapi-cors
fastapi-users[sqlalchemy,oauth]
httpx-oauth
asyncpg
uvicorn
python-dotenv
aiofiles
pyjwt[crypto]
cryptography
apscheduler

# AI & LangChain
langchain
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
langchain_chroma
chromadb
tavily-python

# Automated Testing
pytest
pytest-asyncio
httpx